app/requirements.txt
praw
kafka-python
python-dotenv
pymongo
nltk